#179 use vite instead of static provider
