+++ package.json
@@ -14,50 +14,57 @@
     "test:e2e": "npx playwright test",
     "command": "tsx ./src/cli.ts"
   },
+  "workspaces": [
+    "provider"
+  ],
   "author": "",
   "license": "MIT",
   "dependencies": {
-    "@maplibre/maplibre-gl-style-spec": "^20.3.1",
-    "@types/geojson": "^7946.0.14",
+    "@maplibre/maplibre-gl-style-spec": "^23.1.0",
+    "@types/geojson": "^7946.0.16",
     "@types/jsonminify": "^0.4.3",
     "@unvt/sprite-one": "^0.1.1",
-    "axios": "^1.7.5",
-    "commander": "^12.1.0",
-    "glob": "^11.0.0",
+    "@watergis/maplibre-gl-legend": "^2.0.5",
+    "axios": "^1.8.4",
+    "commander": "^13.1.0",
+    "glob": "^11.0.1",
     "js-yaml": "^4.1.0",
     "jsonminify": "^0.4.2",
+    "maplibre-gl": "^5.3.0",
     "node-watch": "^0.7.4",
     "open": "^10.1.0",
+    "pmtiles": "^4.3.0",
+    "vite": "^6.2.5",
     "ws": "^8.18.1"
   },
   "devDependencies": {
-    "@eslint/eslintrc": "^3.1.0",
-    "@eslint/js": "^9.9.1",
-    "@playwright/test": "^1.46.1",
-    "@types/chai": "^4.3.19",
-    "@types/chai-as-promised": "^7.1.8",
+    "@eslint/eslintrc": "^3.3.1",
+    "@eslint/js": "^9.24.0",
+    "@playwright/test": "^1.15.1",
+    "@types/chai": "^5.2.1",
+    "@types/chai-as-promised": "^8.0.2",
     "@types/fs-extra": "^11.0.4",
     "@types/js-yaml": "^4.0.9",
     "@types/mapbox__point-geometry": "^0.1.4",
-    "@types/mocha": "^10.0.7",
-    "@types/node": "^22.5.0",
-    "@types/ws": "^8.18.0",
-    "@typescript-eslint/eslint-plugin": "^8.3.0",
-    "@typescript-eslint/parser": "^8.3.0",
-    "chai": "^5.1.1",
-    "chai-as-promised": "^8.0.0",
-    "eslint": "^9.9.1",
-    "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-prettier": "^5.2.1",
-    "fs-extra": "^11.2.0",
+    "@types/mocha": "^10.0.10",
+    "@types/node": "^22.14.0",
+    "@types/ws": "^8.18.1",
+    "@typescript-eslint/eslint-plugin": "^8.29.1",
+    "@typescript-eslint/parser": "^8.29.1",
+    "chai": "^5.2.0",
+    "chai-as-promised": "^8.0.1",
+    "eslint": "^9.24.0",
+    "eslint-config-prettier": "^10.1.1",
+    "eslint-plugin-prettier": "^5.2.6",
+    "fs-extra": "^11.3.0",
     "kill-port-process": "^3.2.1",
-    "mocha": "^10.7.3",
+    "mocha": "^11.1.0",
     "node-abort-controller": "^3.1.1",
-    "prettier": "^3.3.3",
+    "prettier": "^3.5.3",
     "ts-node": "^10.9.2",
     "tsconfig-paths": "^4.2.0",
     "tsx": "^4.19.3",
-    "typescript": "^5.5.4"
+    "typescript": "^5.8.3"
   },
   "type": "module"
 }
