+++ provider/index.html
@@ -1,13 +1,7 @@
 <!DOCTYPE html>
 <html id="charites-default">
   <head>
-    <link rel="stylesheet" href="/app.css" />
-    <link href='https://cdn.jsdelivr.net/npm/maplibre-gl/dist/maplibre-gl.css' rel='stylesheet' />
-    <link href='https://cdn.jsdelivr.net/npm/@watergis/maplibre-gl-legend@2.0.5/dist/maplibre-gl-legend.css' rel='stylesheet' />
     <title>Charites Live Preview</title>
-    <script src='https://cdn.jsdelivr.net/npm/maplibre-gl/dist/maplibre-gl.js'></script>
-    <script src="https://cdn.jsdelivr.net/npm/@watergis/maplibre-gl-legend@2.0.5/dist/maplibre-gl-legend.umd.js"></script>
-    <script src="https://cdn.jsdelivr.net/npm/pmtiles@4.3.0/dist/pmtiles.js"></script>
   </head>
   <body>
     <div class="overlay">
@@ -18,11 +12,6 @@
       <label><input type="checkbox" id="showPadding">show Padding</label>
     </div>
     <div id="map"></div>
-    <script type="text/javascript">
-      let protocol = new pmtiles.Protocol()
-        maplibregl.addProtocol('pmtiles', protocol.tile)
-      </script>
-    <script type="text/javascript" src="/shared.js"></script>
-    <script type="text/javascript" src="/app.js"></script>
+    <script type="module" src="./src/main.ts"></script>
   </body>
 </html>
