+++ src/cli/serve.ts
@@ -1,4 +1,7 @@
 import { Command } from 'commander'
+import { createServer, InlineConfig } from 'vite'
+import { fileURLToPath } from 'url'
+import path from 'path'
 import { serve, serveOptions } from '../commands/serve.js'
 import { error } from '../lib/error.js'
 
@@ -13,6 +16,7 @@ program
   )
   .option('--sdf', 'Allows to use SDF sprite in charites')
   .option('--port [port]', 'Specify custom port')
+  .option('--vite-port [port]', 'Specify custom port for vite server')
   .option('--no-open', "Don't open the preview in the default browser")
   .action(async (source: string, serveOptions: serveOptions) => {
     const options: serveOptions = program.opts()
@@ -20,11 +24,26 @@ program
     options.spriteInput = serveOptions.spriteInput
     options.open = serveOptions.open
     options.sdf = serveOptions.sdf
+    options.vitePort = serveOptions.vitePort
     try {
+      await startProviderDevServer(options)
       await serve(source, program.opts())
     } catch (e) {
       error(e)
     }
   })
 
+const __filename = fileURLToPath(import.meta.url)
+const __dirname = path.dirname(__filename)
+async function startProviderDevServer(options: serveOptions) {
+  const config: InlineConfig = {
+    root: path.resolve(__dirname, '../../provider'),
+    server: {
+      port: parseInt(options.vitePort ?? '5137'),
+    },
+  }
+
+  const viteServer = await createServer(config)
+  await viteServer.listen()
+}
 export default program
